requirements.txt
langchain
langchain-community
langchain-openai
openai
google-genai
sentence-transformers
rank-bm25
datasets
faiss-cpu
scikit-learn
numpy
pandas
tqdm
python-dotenv
pytest
jupyter
matplotlib
transformers
bitsandbytes
accelerate
